# openapi/Order workflow API.1.0.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/order/complete": {
      "post": {
        "description": "Confirm completion an existing order based on the provided order ID - done by commissioner.",
        "operationId": "OrderComfirmCompletionController_postOrderCancel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderConfirmCompletionDtoV1"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order canceled successfully"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Order not found (NOT_FOUND)"
          },
          "409": {
            "description": "Illegal state transition (ILLEGAL_TRANSITION)"
          }
        },
        "security": [
          {
            "JWT": []
          },
          {
            "JWT": []
          }
        ],
        "summary": "Confirm order completion",
        "tags": [
          "Order workflow"
        ]
      }
    },
...
  },
  "info": {
    "title": "Order workflow API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "Paste: Bearer <your-JWT>"
      }
    },
    "schemas": {
...
      "OrderConfirmCompletionDtoV1": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "description": "Unique identifier of the order",
            "format": "uuid"
          },
          "workshopId": {
            "type": "string",
            "description": "Unique identifier of the workshop",
            "format": "uuid"
          },
          "commissionerId": {
            "type": "string",
            "description": "Unique identifier of the commissioner",
            "format": "uuid"
          }
        },
        "required": [
          "orderId",
          "workshopId",
          "commissionerId"
        ]
      },
...
    }
  }
}